Hisilicon/D05: update distance of Slit table
D05 platform numa nodes distance:

--------------        --------------
|Node1--Node0|<------>|Node2--Node3|
--------------        --------------
    cluster0              cluster1

Each cluster has two nodes, we treat that the node0
and node1 have the same distance to Node2 and node3,
on the other hand, the node2 and node3 have the same
distance to node0 and node1.
The default distance of Slit table is wrong, correct it.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Chenhui Sun <sunchenhui@huawei.com>